Moved tests to new folder structure.
