+++ areas/monitor/tests/AzureMcp.Monitor.UnitTests/HealthModels/Entity/EntityGetHealthCommandTests.cs
@@ -9,7 +9,6 @@ using AzureMcp.Core.Models.Command;
 using AzureMcp.Core.Options;
 using AzureMcp.Monitor.Commands.HealthModels.Entity;
 using AzureMcp.Monitor.Services;
-using AzureMcp.Tests;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using ModelContextProtocol.Server;
