+++ areas/monitor/tests/AzureMcp.Monitor.UnitTests/Log/ResourceLogQueryCommandTests.cs
@@ -4,40 +4,39 @@
 using System.CommandLine;
 using System.CommandLine.Parsing;
 using System.Text.Json.Nodes;
-using AzureMcp.Areas.Monitor.Commands.Log;
-using AzureMcp.Areas.Monitor.Services;
-using AzureMcp.Models.Command;
-using AzureMcp.Options;
+using AzureMcp.Core.Models.Command;
+using AzureMcp.Core.Options;
+using AzureMcp.Monitor.Commands.Log;
+using AzureMcp.Monitor.Services;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 using Xunit;
 
-namespace AzureMcp.Tests.Areas.Monitor.UnitTests.Log;
+namespace AzureMcp.Monitor.UnitTests.Log;
 
 [Trait("Area", "Monitor")]
-public sealed class WorkspaceLogQueryCommandTests
+public sealed class ResourceLogQueryCommandTests
 {
     private readonly IServiceProvider _serviceProvider;
     private readonly IMonitorService _monitorService;
-    private readonly ILogger<WorkspaceLogQueryCommand> _logger;
-    private readonly WorkspaceLogQueryCommand _command;
+    private readonly ILogger<ResourceLogQueryCommand> _logger;
+    private readonly ResourceLogQueryCommand _command;
     private readonly CommandContext _context;
     private readonly Parser _parser;
 
     private const string _knownSubscription = "knownSubscription";
-    private const string _knownWorkspace = "knownWorkspace";
-    private const string _knownResourceGroup = "knownResourceGroup";
-    private const string _knownTableName = "knownTableName";
+    private const string _knownResourceId = "/subscriptions/sub123/resourceGroups/rg1/providers/Microsoft.Storage/storageAccounts/storage1";
+    private const string _knownTableName = "StorageEvents";
+    private const string _knownQuery = "| limit 10";
     private const string _knownTenant = "knownTenant";
     private const string _knownHours = "24";
     private const string _knownLimit = "100";
-    private const string _knownQuery = "| limit 10";
 
-    public WorkspaceLogQueryCommandTests()
+    public ResourceLogQueryCommandTests()
     {
         _monitorService = Substitute.For<IMonitorService>();
-        _logger = Substitute.For<ILogger<WorkspaceLogQueryCommand>>();
+        _logger = Substitute.For<ILogger<ResourceLogQueryCommand>>();
 
         var collection = new ServiceCollection();
         collection.AddSingleton(_monitorService);
@@ -49,9 +48,9 @@ public sealed class WorkspaceLogQueryCommandTests
     }
 
     [Theory]
-    [InlineData($"--subscription {_knownSubscription} --workspace {_knownWorkspace} --resource-group {_knownResourceGroup} --table-name {_knownTableName} --query {_knownQuery}", true)]
-    [InlineData($"--subscription {_knownSubscription} --workspace {_knownWorkspace} --resource-group {_knownResourceGroup} --table-name {_knownTableName} --query {_knownQuery} --hours {_knownHours} --limit {_knownLimit}", true)]
-    [InlineData($"--subscription {_knownSubscription} --workspace {_knownWorkspace} --table-name {_knownTableName} --query {_knownQuery}", false)] // missing resource-group
+    [InlineData($"--subscription {_knownSubscription} --resource-id {_knownResourceId} --table-name {_knownTableName} --query {_knownQuery}", true)]
+    [InlineData($"--subscription {_knownSubscription} --resource-id {_knownResourceId} --table-name {_knownTableName} --query {_knownQuery} --hours {_knownHours} --limit {_knownLimit}", true)]
+    [InlineData($"--subscription {_knownSubscription} --table-name {_knownTableName} --query {_knownQuery}", false)] // missing resource-id
     [InlineData($"--subscription {_knownSubscription}", false)]
     [InlineData("", false)]
     public async Task ExecuteAsync_ValidatesInputCorrectly(string args, bool shouldSucceed)
@@ -61,14 +60,14 @@ public sealed class WorkspaceLogQueryCommandTests
         {
             var mockResults = new List<JsonNode>
             {
-                JsonNode.Parse(@"{""TimeGenerated"": ""2023-01-01T12:00:00Z"", ""Message"": ""Test log entry""}") ?? JsonNode.Parse("{}") ?? new JsonObject(),
-                JsonNode.Parse(@"{""TimeGenerated"": ""2023-01-01T12:01:00Z"", ""Message"": ""Another log entry""}") ?? JsonNode.Parse("{}") ?? new JsonObject()
+                JsonNode.Parse(@"{""TimeGenerated"": ""2023-01-01T12:00:00Z"", ""Message"": ""Resource log entry""}") ?? JsonNode.Parse("{}") ?? new JsonObject(),
+                JsonNode.Parse(@"{""TimeGenerated"": ""2023-01-01T12:01:00Z"", ""Message"": ""Another resource log entry""}") ?? JsonNode.Parse("{}") ?? new JsonObject()
             };
-            _monitorService.QueryWorkspaceLogs(
-                Arg.Any<string>(), 
-                Arg.Any<string>(), 
-                Arg.Any<string>(), 
-                Arg.Any<string>(), 
+            _monitorService.QueryResourceLogs(
+                _knownSubscription, 
+                _knownResourceId, 
+                _knownQuery, 
+                _knownTableName, 
                 Arg.Any<int?>(), 
                 Arg.Any<int?>(), 
                 Arg.Any<string>(), 
@@ -98,15 +97,15 @@ public sealed class WorkspaceLogQueryCommandTests
         // Arrange
         var mockResults = new List<JsonNode>
         {
-            JsonNode.Parse(@"{""TimeGenerated"": ""2023-01-01T12:00:00Z"", ""Message"": ""Test log entry"", ""Level"": ""Info""}") ?? new JsonObject(),
-            JsonNode.Parse(@"{""TimeGenerated"": ""2023-01-01T12:01:00Z"", ""Message"": ""Another log entry"", ""Level"": ""Warning""}") ?? new JsonObject(),
-            JsonNode.Parse(@"{""TimeGenerated"": ""2023-01-01T12:02:00Z"", ""Message"": ""Error occurred"", ""Level"": ""Error""}") ?? new JsonObject()
+            JsonNode.Parse($@"{{""TimeGenerated"": ""2023-01-01T12:00:00Z"", ""ResourceId"": ""{_knownResourceId}"", ""Level"": ""Info""}}") ?? new JsonObject(),
+            JsonNode.Parse($@"{{""TimeGenerated"": ""2023-01-01T12:01:00Z"", ""ResourceId"": ""{_knownResourceId}"", ""Level"": ""Warning""}}") ?? new JsonObject(),
+            JsonNode.Parse($@"{{""TimeGenerated"": ""2023-01-01T12:02:00Z"", ""ResourceId"": ""{_knownResourceId}"", ""Level"": ""Error""}}") ?? new JsonObject()
         };
-        _monitorService.QueryWorkspaceLogs(
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
+        _monitorService.QueryResourceLogs(
+            _knownSubscription, 
+            _knownResourceId, 
+            _knownQuery, 
+            _knownTableName, 
             Arg.Any<int?>(), 
             Arg.Any<int?>(), 
             Arg.Any<string>(), 
@@ -115,8 +114,7 @@ public sealed class WorkspaceLogQueryCommandTests
 
         var args = _parser.Parse([
             "--subscription", _knownSubscription,
-            "--workspace", _knownWorkspace,
-            "--resource-group", _knownResourceGroup,
+            "--resource-id", _knownResourceId,
             "--table-name", _knownTableName,
             "--query", _knownQuery
         ]);
@@ -129,11 +127,11 @@ public sealed class WorkspaceLogQueryCommandTests
         Assert.NotNull(response.Results);
 
         // Verify the mock was called
-        await _monitorService.Received(1).QueryWorkspaceLogs(
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
+        await _monitorService.Received(1).QueryResourceLogs(
+            _knownSubscription, 
+            _knownResourceId, 
+            _knownQuery, 
+            _knownTableName, 
             Arg.Any<int?>(), 
             Arg.Any<int?>(), 
             Arg.Any<string>(), 
@@ -145,9 +143,9 @@ public sealed class WorkspaceLogQueryCommandTests
     {
         // Arrange
         var mockResults = new List<JsonNode> { JsonNode.Parse(@"{""result"": ""data""}") ?? new JsonObject() };
-        _monitorService.QueryWorkspaceLogs(
+        _monitorService.QueryResourceLogs(
             _knownSubscription,
-            _knownWorkspace,
+            _knownResourceId,
             _knownQuery,
             _knownTableName,
             int.Parse(_knownHours),
@@ -158,8 +156,7 @@ public sealed class WorkspaceLogQueryCommandTests
 
         var args = _parser.Parse([
             "--subscription", _knownSubscription,
-            "--workspace", _knownWorkspace,
-            "--resource-group", _knownResourceGroup,
+            "--resource-id", _knownResourceId,
             "--table-name", _knownTableName,
             "--query", _knownQuery,
             "--hours", _knownHours,
@@ -172,9 +169,9 @@ public sealed class WorkspaceLogQueryCommandTests
 
         // Assert
         Assert.Equal(200, response.Status);
-        await _monitorService.Received(1).QueryWorkspaceLogs(
+        await _monitorService.Received(1).QueryResourceLogs(
             _knownSubscription,
-            _knownWorkspace,
+            _knownResourceId,
             _knownQuery,
             _knownTableName,
             int.Parse(_knownHours),
@@ -188,11 +185,11 @@ public sealed class WorkspaceLogQueryCommandTests
     {
         // Arrange
         var mockResults = new List<JsonNode> { JsonNode.Parse(@"{""result"": ""data""}") ?? new JsonObject() };
-        _monitorService.QueryWorkspaceLogs(
-            Arg.Any<string>(),
-            Arg.Any<string>(),
-            Arg.Any<string>(),
-            Arg.Any<string>(),
+        _monitorService.QueryResourceLogs(
+            _knownSubscription,
+            _knownResourceId,
+            _knownQuery,
+            _knownTableName,
             Arg.Any<int?>(),
             Arg.Any<int?>(),
             Arg.Any<string>(),
@@ -201,8 +198,7 @@ public sealed class WorkspaceLogQueryCommandTests
 
         var args = _parser.Parse([
             "--subscription", _knownSubscription,
-            "--workspace", _knownWorkspace,
-            "--resource-group", _knownResourceGroup,
+            "--resource-id", _knownResourceId,
             "--table-name", _knownTableName,
             "--query", _knownQuery
         ]);
@@ -212,9 +208,9 @@ public sealed class WorkspaceLogQueryCommandTests
 
         // Assert
         Assert.Equal(200, response.Status);
-        await _monitorService.Received(1).QueryWorkspaceLogs(
+        await _monitorService.Received(1).QueryResourceLogs(
             _knownSubscription,
-            _knownWorkspace,
+            _knownResourceId,
             _knownQuery,
             _knownTableName,
             Arg.Any<int?>(), // Default hours
@@ -227,11 +223,11 @@ public sealed class WorkspaceLogQueryCommandTests
     public async Task ExecuteAsync_HandlesServiceErrors()
     {
         // Arrange
-        _monitorService.QueryWorkspaceLogs(
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
-            Arg.Any<string>(), 
+        _monitorService.QueryResourceLogs(
+            _knownSubscription, 
+            _knownResourceId, 
+            _knownQuery, 
+            _knownTableName, 
             Arg.Any<int?>(), 
             Arg.Any<int?>(), 
             Arg.Any<string>(), 
@@ -240,8 +236,7 @@ public sealed class WorkspaceLogQueryCommandTests
 
         var args = _parser.Parse([
             "--subscription", _knownSubscription,
-            "--workspace", _knownWorkspace,
-            "--resource-group", _knownResourceGroup,
+            "--resource-id", _knownResourceId,
             "--table-name", _knownTableName,
             "--query", _knownQuery
         ]);
@@ -253,5 +248,47 @@ public sealed class WorkspaceLogQueryCommandTests
         Assert.Equal(500, response.Status);
         Assert.Contains("Test error", response.Message);
         Assert.Contains("troubleshooting", response.Message);
+    }
+
+    [Fact]
+    public async Task ExecuteAsync_WithComplexResourceId_HandlesCorrectly()
+    {
+        // Arrange
+        var complexResourceId = "/subscriptions/12345678-1234-1234-1234-123456789012/resourceGroups/my-rg/providers/Microsoft.Compute/virtualMachines/my-vm";
+        var query = "| where Level == 'Error'";
+        var table = "VMEvents";
+        var mockResults = new List<JsonNode> { JsonNode.Parse(@"{""result"": ""vm data""}") ?? new JsonObject() };
+        _monitorService.QueryResourceLogs(
+            _knownSubscription,
+            complexResourceId,
+            query,
+            table,
+            Arg.Any<int?>(),
+            Arg.Any<int?>(),
+            Arg.Any<string>(),
+            Arg.Any<RetryPolicyOptions>())
+            .Returns(mockResults);
+
+        var args = _parser.Parse([
+            "--subscription", _knownSubscription,
+            "--resource-id", complexResourceId,
+            "--table-name", table,
+            "--query", query
+        ]);
+
+        // Act
+        var response = await _command.ExecuteAsync(_context, args);
+
+        // Assert
+        Assert.Equal(200, response.Status);
+        await _monitorService.Received(1).QueryResourceLogs(
+            _knownSubscription,
+            complexResourceId,
+            query,
+            table,
+            Arg.Any<int?>(),
+            Arg.Any<int?>(),
+            Arg.Any<string>(),
+            Arg.Any<RetryPolicyOptions>());
     }
 }
