+++ areas/monitor/tests/AzureMcp.Monitor.UnitTests/Metrics/MetricsDefinitionsCommandTests.cs
@@ -8,7 +8,6 @@ using AzureMcp.Core.Options;
 using AzureMcp.Monitor.Commands.Metrics;
 using AzureMcp.Monitor.Models;
 using AzureMcp.Monitor.Services;
-using AzureMcp.Tests;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
